requirements.txt
pyrocketbot
rocketchat_API
rocket-python
RocketChatAPIBot
